\name{greetr-package}
\alias{greetr-package}
\alias{greetr}
\docType{package}
\title{
\packageTitle{greetr}
}
\description{
\packageDescription{greetr}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{greetr}

Maintainer: \packageMaintainer{greetr}
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = greetr)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{greetr}}
%\section{Documentation Index}{\packageIndices{greetr}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
