\name{hello}
\alias{hello}
\title{Say Hello}
\description{
The \code{hello} function takes a string \code{x} and prints a friendly greeting:
\code{"Hello x!"} with the value of \code{x} substituted in.
}
\usage{
hello(x)
}
\arguments{
  \item{x}{
A character string giving the name or text to include in the greeting.
}
}
\value{
This function prints a message to the console. It does not return a value.
}
\author{
Emma Mott
}
\examples{
hello("world")
}
